Default config URL to app's PlayStore URL
For flavoured builds, the URL for the Update button will differ and make
it necessary to manage the JSON and request configuration in multiple
JSON files.

As a sane default, this commit adds the package's Play Store URL in case
no URL is provided in the config.json. This means the change is fully
backwards-compatible.