Prevent exception when activity is destroyed
When a user taps the back button or otherwise triggers the activity to
be destroyed, this would lead to an exception in Glider, force-closing
the app with:

```
Fatal Exception: java.lang.IllegalArgumentException: You cannot start a load for a destroyed activity
       at com.bumptech.glide.manager.RequestManagerRetriever.assertNotDestroyed(RequestManagerRetriever.java:317)
       at com.bumptech.glide.manager.RequestManagerRetriever.get(RequestManagerRetriever.java:128)
       at com.bumptech.glide.Glide.with(Glide.java:801)
...
```

When the activity is destroyed, it's pointless trying to load the
resource anyway, so this commit wraps calls to Glider with an according
check. The solution was taken from a discussion[1] on GitHub.

[1] https://github.com/bumptech/glide/issues/1484#issuecomment-365625087