Support dynamic configuration of JSON config links
Currently the URLs (or packages) of wallpaper_json and config_json are
located in XML string resources. As these can't be overridden in Java,
there is no way to dynamically configure them to run e.g. A/B tests with
multiple or user-specific configurations.

This commit refactors all usages of the string resources to using a
newly introduced ConfigHandler. Fully backwards-compatible because the
handler defaults back to the string resources.

Fully configurable because for retrieving the URLs at runtime a remote
service such as Firebase Remote Config could be used.