Allow overriding TapIntro colours
This commit adds six additional parameters to configure:

Colors:
 * `tapIntroTitle`
 * `tapIntroDescription`
 * `tapIntroCircleInner`
 * `tapIntroCircleOuter`

Boolean:
 * `use_legacy_intro_colors` (defaults to `true`)

Integer:
 * `tap_intro_circle_scale_percent` (defaults to 100%)

The default behaviour for CandyBar users does not change. Those who
want to customise the tap intro colours, can turn the default behaviour
off with `use_legacy_intro_colors=false` and then go on to specify
custom colours with the 4 color values.

Also, a scale factor has been added for the tap intro circles. It will
be taken as-is and will multiply the default values accordingly to
increase or decrease the size of the outer circle.