Add support for visibly disabled requests
This commit follows the discussion in
https://github.com/zixpo/candybar/discussions/133 and makes more options
available to the FilterRequestHandler. In addition to the return value
of true/false which completely hides apps from the icon request section,
the new methods `setAvailableForRequest()` and `setInfoText()` allow to
arbitrarily annotate and disable icons as follows:

When `setInfoText()` sets a non-empty string, a little `(i)` icon will
be displayed and upon tapping it, a dialog with `infoText` pops up.

When `setAvailableForRequest()` is set to `false`, the entire item will
be greyed out. When tapping it, nothing happens. Unless `setInfoText()`
is used as well, in which case a dialog with the `infoText` pops up.

This allows for flexible combinations of selectively hiding, greying
out, or just annotating icons based on matched patterns.