Control icon reshaping in Nova launcher
This commit adds a new configuration option for using "direct apply" in
Nova launcher. For icon packs with intentionally transparent
backgrounds, Nova unfortunately defaults to its "Reshape legacy icons"
setting being "ON" and adding backgrounds. For monochrome packs, it
works ok even if unexpected for users. For colourful packs, the setting
often guesses the background colours very wrong leading to ugly and
illegible results.

The Nova team agreed on exposing this feature to devs via an Intent
Extra. The default behaviour "ON" is kept but can be overridden by the
icon pack to "OFF" (disabling it for all icons) or "AUTO" (disabling it
for unthemed icons only).

Refs https://discord.com/channels/466300922767736844/1098328998532096142/1109611038401368084