Refactor into ColorOS and OxygenOS launchers
What a long story. At some point, ColorOS and OxygenOS were completely
separate launchers and tied to their respective devices. As of Android
12, both are starting to merge and no longer reporting as
`com.oppo.launcher` and `net.oneplus.launcher`, but instead as
`com.android.launcher` which is just identifying the Stock Android
launcher.

However, their variants are very different from Stock Android. The good
news is that their new merge supports icon packs (so it's worth the
effort) but the bad news is that as of August 2023 it's unclear how
their merge will proceed further beyond Android 13/14.

See the original PR description for a full list of tested devices.