Add support for re-requesting icons
This commit changes the default behaviour and adds new behaviour hidden
behind a new boolean option `enable_icon_request_multiple`.

 * When `enable_icon_request_multiple=false` (default), users cannot
   select icons they already requested but instead see an info dialog.

 * When `enable_icon_request_multiple=true`, users see a confirmation
   dialog for icons they already requested, asking them whether they are
   sure they want to spend another request on them.

The old default behaviour was to let the user select already requested
icons but then display an error dialog when they hit the "send" button.
From a user perspective it never made sense to let them select it in the
first place, as the result was 100% sure to error out.

The new default behaviour happens immediately on click of the checkbox.

The new multi-request behaviour behind the flag is useful for those who
want to support a voting system for free requests (i.e. let users cast
multiple votes for the same icon), and useful for those with both free
and premium requests (e.g. when users want to "top-up" their free
requests later on to speed them up).