Use success callback
The Candybar default behaviour is to close the activity once a
meaningful action has been done, e.g. the icon pack has been applied or
the user tapped on a dialog option. This might not always be desirable,
so we provide a DEFAULT_CALLBACK which preserves what Candybar is doing
but also allows overrides.